# openapi.yaml
openapi: 3.0.0
info:
  title: Riot Stats API
  version: 1.0.0
  description: >
    API REST pour trois collections MongoDB: match_ids, players, match_data.
    Les identifiants sont:
    - match_ids: matchId
    - players: puuid
    - match_data: metadata.matchId
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/match-ids:
    get:
      tags: [match-ids]
      summary: Lister les match_ids
      operationId: listMatchIds
      description: Liste paginée, filtrable par tier et rank.
      parameters:
        - in: query
          name: tier
          schema: { $ref: "#/components/schemas/Tier" }
        - in: query
          name: rank
          schema: { $ref: "#/components/schemas/Rank" }
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: sort
          schema: { type: string, example: "matchId,asc" }
          description: "Champ,ordre (ex. matchId,asc | tier,desc)"
      responses:
        "200":
          description: Page de résultats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageMatchId"
    post:
      tags: [match-ids]
      summary: Créer un matchId
      operationId: createMatchId
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MatchId" }
      responses:
        "201":
          description: Créé
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MatchId" }
        "409":
          description: Conflit (matchId déjà existant)
  /api/match-ids/bulk:
    post:
      tags: [match-ids]
      summary: Upsert en lot (bulk upsert)
      operationId: bulkUpsertMatchIds
      description: Crée ou met à jour une liste de documents (clé = matchId).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/MatchId" }
      responses:
        "200":
          description: Résultat du bulk
          content:
            application/json:
              schema:
                type: object
                properties:
                  inserted: { type: integer }
                  updated: { type: integer }
                  total:   { type: integer }
  /api/match-ids/{matchId}:
    get:
      tags: [match-ids]
      summary: Récupérer par matchId (clé métier)
      operationId: getMatchIdByMatchId
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: string, example: "NA1_5348438296" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MatchId" }
          links:
            matchDataByMatchId:
              operationId: getMatchDataByMatchId
              parameters:
                matchId: "$response.body#/matchId"
        "404":
          description: Introuvable
    put:
      tags: [match-ids]
      summary: Remplacer (PUT) un matchId
      operationId: putMatchIdByMatchId
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MatchId" }
      responses:
        "200":
          description: Mis à jour
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MatchId" }
        "404":
          description: Introuvable
    patch:
      tags: [match-ids]
      summary: Mise à jour partielle (PATCH) par matchId
      operationId: patchMatchIdByMatchId
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                tier: { $ref: "#/components/schemas/Tier" }
                rank: { $ref: "#/components/schemas/Rank" }
      responses:
        "200":
          description: Mis à jour
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MatchId" }
        "404":
          description: Introuvable
    delete:
      tags: [match-ids]
      summary: Supprimer par matchId
      operationId: deleteMatchIdByMatchId
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Supprimé
        "404":
          description: Introuvable
  /api/match-ids/stats/count:
    get:
      tags: [match-ids]
      summary: Compter les documents (option filtres)
      operationId: countMatchIds
      parameters:
        - in: query
          name: tier
          schema: { $ref: "#/components/schemas/Tier" }
        - in: query
          name: rank
          schema: { $ref: "#/components/schemas/Rank" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer, example: 12345 }
...
  /api/match-ids/{matchId}/data:
    get:
      tags: [match-ids, match-data]
      summary: Récupérer le document match_data associé à ce matchId
      operationId: getMatchDataFromMatchId
      parameters:
        - in: path
          name: matchId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MatchData" }
        "404":
          description: Introuvable

  ##########################
  # PLAYERS (key = puuid)
  ##########################
...
components:
  schemas:
    Tier:
      type: string
      description: Palier de league
      enum: [IRON, BRONZE, SILVER, GOLD, PLATINUM, EMERALD, DIAMOND, MASTER, GRANDMASTER, CHALLENGER]
      example: PLATINUM
    Rank:
      type: string
      description: Division au sein du palier
      enum: [I, II, III, IV]
      example: I

    ##########################
    # MATCH IDS SCHEMAS
    ##########################
    MatchId:
      type: object
      additionalProperties: false
      properties:
        matchId:
          type: string
          example: "NA1_5348438296"
        tier:
          $ref: "#/components/schemas/Tier"
        rank:
          $ref: "#/components/schemas/Rank"
      required: [matchId, tier, rank]
    PageMatchId:
      type: object
      properties:
        content:
          type: array
          items: { $ref: "#/components/schemas/MatchId" }
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 50
        totalElements:
          type: integer
          example: 1234
        totalPages:
          type: integer
          example: 25
        sort:
          type: string
          example: "matchId,asc"

    ##########################
    # PLAYERS SCHEMAS
    ##########################
...
    MatchDataMetadata:
      type: object
      additionalProperties: false
      properties:
        dataVersion: { type: string, example: "2" }
        matchId: { type: string, example: "NA1_5348438296" }
        participants:
          type: array
          items: { type: string, description: "PUUID" }
      required: [dataVersion, matchId, participants]
    MatchDataTeamObjectives:
      type: object
      additionalProperties: true
      properties:
        first: { type: boolean }
        kills: { type: integer }
    MatchDataTeamFeats:
      type: object
      additionalProperties: true
      description: "Riot feats map, ex: FIRST_BLOOD, FIRST_TURRET…"
      properties:
        FIRST_BLOOD:
          type: object
          properties:
            featState: { type: integer }
    MatchDataTeam:
      type: object
      additionalProperties: false
      properties:
        teamId: { type: integer, enum: [100, 200] }
        win: { type: boolean }
        bans:
          type: array
          items:
            type: object
            properties:
              championId: { type: integer }
              pickTurn: { type: integer }
        feats:
          $ref: "#/components/schemas/MatchDataTeamFeats"
        objectives:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/MatchDataTeamObjectives"
    MatchDataParticipant:
      type: object
      description: "Participant unique d’une partie"
      additionalProperties: true
      properties:
        puuid: { type: string }
        participantId: { type: integer, minimum: 1, maximum: 10 }
        teamId: { type: integer, enum: [100, 200] }
        championId: { type: integer }
        championName: { type: string }
        individualPosition: { type: string, example: "TOP" }
        teamPosition: { type: string, example: "TOP" }
        kills: { type: integer }
        deaths: { type: integer }
        assists: { type: integer }
        win: { type: boolean }
        goldEarned: { type: integer }
        totalDamageDealtToChampions: { type: integer }
        totalMinionsKilled: { type: integer }
        summonerName: { type: string }
        summonerId: { type: string }
        riotIdGameName: { type: string }
        riotIdTagline: { type: string }
        itemsPurchased: { type: integer }
        item0: { type: integer }
        item1: { type: integer }
        item2: { type: integer }
        item3: { type: integer }
        item4: { type: integer }
        item5: { type: integer }
        item6: { type: integer }
        perks:
          type: object
          additionalProperties: true
    MatchDataInfo:
      type: object
      additionalProperties: false
      properties:
        endOfGameResult: { type: string, example: "GameComplete" }
        gameCreation: { type: integer, format: int64, description: "epoch ms" }
        gameDuration: { type: integer, description: "seconds" }
        gameEndTimestamp: { type: integer, format: int64, description: "epoch ms" }
        gameId: { type: integer, format: int64 }
        gameMode: { type: string, example: "CLASSIC" }
        gameName: { type: string }
        gameStartTimestamp: { type: integer, format: int64, description: "epoch ms" }
        gameType: { type: string, example: "MATCHED_GAME" }
        gameVersion: { type: string, example: "15.16.704.6097" }
        mapId: { type: integer, example: 11 }
        platformId: { type: string, example: "NA1" }
        queueId: { type: integer, example: 440, description: "e.g., Flex 5v5" }
        participants:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/MatchDataParticipant" }
        teams:
          type: array
          items: { $ref: "#/components/schemas/MatchDataTeam" }
        tournamentCode: { type: string }
    MatchData:
      type: object
      additionalProperties: false
      properties:
        metadata: { $ref: "#/components/schemas/MatchDataMetadata" }
        info: { $ref: "#/components/schemas/MatchDataInfo" }
      required: [metadata, info]
...